% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suburbs_crime.R
\name{suburbs_crime}
\alias{suburbs_crime}
\title{HW2_function
\code{suburbs_crime} The fucntion take the user inputs, including two surburs,
offence_descritpyions, and plot the correlation in offence count between the two suburbs}
\usage{
suburbs_crime(crime_data, offence_description, suburbs)
}
\arguments{
\item{crime_data}{A data.table object with the following columns:
"date" (POSIXct), "suburb" (chr), "postcode" (chr), "offence_level_1" (chr),
"offence_level_2" (chr), "offence_level_3" (chr), "offence_count" (num).}

\item{offence_description}{A character string of offence level 3.}

\item{suburbs}{A two-element character vector. Each element is the name (UPPERCASE)
of an SA suburb.}
}
\value{
A ggplot object showing the correlation in offence count between the two input suburbs.
}
\description{
HW2_function
\code{suburbs_crime} The fucntion take the user inputs, including two surburs,
offence_descritpyions, and plot the correlation in offence count between the two suburbs
}
\examples{
suburbs_crime(datatable,"OFFENCES AGAINST PROPERTY", c("WEST BEACH", "ADELAIDE AIRPORT"))
}
